@@ -4,7 +4,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -27,6 +27,7 @@ flutter:
   uses-material-design: true
 dependency_overrides:
   global_repository:
-    path: /Users/nightmare/Desktop/nightmare-space/global_repository
+    path: ../global_repository
   file_manager_view:
-    path: /Users/nightmare/Desktop/nightmare-space/file_manager_view
+    path: ../file_manager_view
+  flutter_svg: ^1.1.6
