@@ -4,11 +4,11 @@ import 'directory_select_page.dart';
 import 'file_select_page.dart';
 
 class FileSelector {
-  static Future<List<String>> pick(
+  static Future<List<String>?> pick(
     BuildContext context, {
-    String path,
+    String? path,
   }) async {
-    List<String> result = await Navigator.of(context).push(
+    List<String>? result = await Navigator.of(context).push(
       MaterialPageRoute(
         builder: (_) {
           return Theme(
@@ -23,8 +23,8 @@ class FileSelector {
     return result;
   }
 
-  static Future<String> pickDirectory(BuildContext context) async {
-    String result = await Navigator.of(context).push(
+  static Future<String?> pickDirectory(BuildContext context) async {
+    String? result = await Navigator.of(context).push(
       MaterialPageRoute(
         builder: (_) {
           return Theme(
